add condition to handle no posts templates

+++ post/post.service.js
@@ -13,42 +13,72 @@ module.exports = {
 };
 
 async function getAll () {
+  if ( !CPost ) {
+    return { message: 'No new posts exists' };
+  }
+
   // TO DO: GET ONLY SUMMARY
-    return await CPost.find({ type: 'post' }).select('-hash');
+  return await CPost.find({ type: 'post' }).select('-hash');
 }
 
 async function getById (id) {
-    return await CPost.findById(id).select('-hash');
+  if ( !CPost ) {
+    return { message: 'No posts exists' };
+  }
+
+  return await CPost.findById(id).select('-hash');
 }
 
 async function getArchive () {
+  if ( !CPost ) {
+    return { message: 'No posts exists' };
+  }
+
   return await CPost.find({type: 'archive'}).select('-hash');
 }
 
 async function getDataPost () {
+  if ( !CPost ) {
+    return { message: 'No posts exists' };
+  }
+
   return await CPost.find({type: 'data'}).select('-hash');
 }
 
 async function getAllActivePosts () {
+  if ( !CPost ) {
+    return { message: 'No posts exists' };
+  }
+
   return await CPost.find({$nor: {type: 'archive'} }).select('-hash');
 }
 
 async function createPost (inpParam) {
-    const CPost = new CPost(inpParam);
+  const CPost = new CPost(inpParam);
 
-    // save post
-    await CPost.save();
+  // save post
+  await CPost.save();
 }
 
 async function update(id, inpParam) {
-    const CPost = await CPost.findById(id);
+  if ( !CPost ) {
+    return { message: 'No posts exists' };
+  }
 
-    // copy inpParam properties to CPost
-    Object.assign(CPost, inpParam);
+  const CPost = await CPost.findById(id);
 
-    await CPost.save();
+  // copy inpParam properties to CPost
+  Object.assign(CPost, inpParam);
+
+  await CPost.save();
+  return { message: 'Post save successful' };
 }
 
 async function _delete(id) {
-    await CPost.findByIdAndRemove(id);
+  if ( !CPost ) {
+    return { message: 'No posts exists' };
+  }
+
+  await CPost.findByIdAndRemove(id);
+  return { message: 'Post delete successful' };
 }
